ECO-871. Move shipment form to another location.
